[REFACTOR] Created new panel-preview + scene now consistent with hubble's scene
